requirements.txt
docker
flask
flask-socketio
pymysql
DBUtils
python-dotenv
Flask-SQLAlchemy
flask-migrate
alembic